phc_microservice_tbsurveillance/src/main/java/com/ssf/tbsurveillance/config/AuditAwareImpl.java
package com.ssf.tbsurveillance.config;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.AuditorAware;

import javax.servlet.http.HttpServletRequest;
import java.util.Optional;
import com.ssf.tbsurveillance.constants.Constants;
import com.ssf.tbsurveillance.utils.Utils;

public class AuditAwareImpl implements AuditorAware<String> {

    @Autowired
    HttpServletRequest request;

    @Override
    public Optional<String> getCurrentAuditor() {
        String idToken = request.getHeader(Constants.XUSER_ID_HEADER);
        return Optional.of(Utils.getUserIdFromIdToken(idToken));
    }
}
